PropertyContainer: aParentDescriptor is now NULL whenever a property access path enters a new C++ container object.
Until now, is was NULL only for the DsAddressable level.

aParentDescriptor is meant for the special case of several property levels managed by the same C++ object - and makes little sense passing cross-object.

Therefore, aParentDescriptor==NULL can be used to check for being at the first level of properties in a C++ object in general, not only DsAddressable.

The change required some more checks for aParentDescriptor==NULL, some of them using the new HAS_OKEY() macro.