dali: implemented device info cache, to speed up group forming/deletion
before, after defining or deleting a group or composite device, the entire bus with all device infos was re-scanned. Now, no rescanning is needed any more, because the info from the cache can be used.

The busdevice's deviceInfo is now a pointer, so the actual deviceInfo entries still exist only once, but can be referenced from the cache map and the busdevice.