vdchost: improved behaviour in case some vdcs return error when initializing.
This will not abort initializing other vdcs any more, but
will set the affected vdc's vdcErr, and will block devices from
being collected for this vdc.

After initial vdchost-initiated collecting, if any of the vdcs has a non-OK getVdcStatus(), this is returned to the caller (which then can e.g. turn on an error LED)