Vdc: made device identification call chain more safe regarding identified Device object's lifetime.
Now the initial trigger's call chain that causes initialiation to fail is unwound completely before non-identificed Device object is deleted.

Before, the deletion was implicitly happening *somewhere* in the process of unwinding the trigger's call chain, which could lead to device object's memory being accessed after deletion and thus memory corruption (because not all callbacks are in strict tail position).