Memory: fixed possible memory corruption - P44Objs instantiated as member variables must be flagged as such using isMemberVariable() now.
Background: P44Objs as member variables were safe as long as the object was not ever referenced by a intrusivePtr, and thus these worked fine for a long time. However, when referenced by a Ptr, that Ptr going out of scope would decrement the refcount to zero and thus free() the object which was never allocated separately.
Only a recent change in p44utils (01d4920af2 - OperationQueue: aborting operations...) which took a intrusivePtr reference of the queue for keepalive reasons has now surfaced the problem.