SingleDevice: added "standardActions" (which are like customActions, but not API/user changeable)
The common base class ActionMacro refers to a deviceAction and contains some parameters.
This base class has now two subclasses:

- CustomAction is a macro that can be created/edited/deleted via vDC API
  (and eventually, by the user)
- StandardAction is a macro which is defined in vdc-side static configuration
  and can only be viewed/called via the vDC API

StandardActions can be created programmatically in device implementations
or can be loaded as part of a config file using standardActionsFromJSON().
As a shortcut, "autoAddStandardActions" boolean config option or
autoAddStandardActions() method creates a 1:1 standardAction for every
deviceAction.