Color Light Devices: transition stepping was not entirely correct, first step was applying starting value rather than actual first step of change. So there was always a second step applying the end value, even for instant transitions.
With the change, the first step applied is a real step away from the current value. For immediate transitions, it is the only step.

For fast steppers as used for DMX/OLA, LED chains and analog PWMs, this does not make a noticeable difference. For slow steppers where the actual hardware device can do some smoothing of its own, like DALI, the chance is important.