DALI: blame devices with no proper unique Serial as only 90% good in the opState level.
Background: I am slowly loosing temper with all those s****y DALI devices featuring an expensive DALI logo but simply NOT BEING CONFORMANT with the specs by just delivering garbage instead of a proper unique serial number! In the age of IoT, the importance of having unique ids should really be clear to every last firmware developer on the planet. And IT IS IN THE SPECS for probably a decade!