native scenes/groups: basic skeleton for hashing affected devices and scene contents, but no actual hash matching & vdc level operations.
Compiles, quickly tested to work normally for callScene and dimChannel (including lecagy dimming with INC/DEC/STOP callScene).

The decision for optimization is on the device level via prepareForOptimizedSet().
This allows to have only some devices of a vdc participate in an optimized set.
For example, DALI devices already formed by a DALI group cannot be optimized.