Device: area dimming: dimming two areas in succession did continue dimming the first along with the second
Reason was that the last area being dimmed (areaDimmed, for T1234_CONT) was not updated for devices excluded from the current dimming (by dontCare). Therefore, the area of the first dim operation of a device in an area was retained, such that a T1234_CONT for a subsequent dimming in a *different* area would still be interpreted as a continuation of the first, causing multiple areas to dim together.