scene optimizer: fix for dimming: delivery state must not be kept alive longer than actual execution of the notification.
Problem was that the action repeater (for autostopping dimming) was keeping a reference to the original notification, thus keeping the object alive and vdc in "delivering" state. So subsequent inc/dec or dimChannel retriggers where held in the delivery queue until the repeater fired, messing up dimming.

Now the action repeater uses a (shallow) copy of the original delivery state to make sure the notification completes immediately after starting/stopping/retriggering the dimming.