ClimateControlBehaviour: FCU now implements powerState_forcedOff and auto-wake for building protection
When the FCU is set into powerState==2 (powerState_forcedOff), either by the device itself
(which then must report the change in the powerState channel) or by explicitly changing
the channel value via vDC API, all further scene calls are suppressed.

The FCU is re-activated automatically when:
- climate control is in a "protective heating" mode (scene 3 or 5) and TemperatureZone falls below TemperatureSetPoint
- or climate control is in a "protective cooling" mode (scene 12 or 14) and TemperatureZone rises above TemperatureSetPoint

In these cases, the respective "protective" scene is automatically applied to the FCU device,
which should switch it on and also put it into the correct operation mode.