Enocean: Weather station: clarify channel names, change former unspecified "daylight" binary to "twilight" (inverted polarity)
We had an inverted version of "twilight" so far, because Eltako did not document that bit properly for the FWS61, so it was a bit unclear. However original Elsner docs are clear, and the direct P03/04 RS485 weather station external device always had "twilight" for that, so the change makes both consistent now.