Hue: add workaround for broken third-party (TCI, not hue certified) lights which produce random output when receiving "on" and "bri" in the same command, and also when transition time changes between two calls.
Solution for these devices is now to send "on" separately beforehand, when change of "bri" is also needed, plus some optimisations when light is just turned on and off w/o change of "bri".

The re-apply mechanism now uses the same transition time as the original command.

Also, the option for continuous re-apply is removed. This was once intended as a fix for occasionally resetting-to-white hues, but never actually used.
Today, an affected bulb, updated to latest firmware, can re-apply the last color/brightness before the reset itself.