channelbehaviour: make sure incomplete channel transitions are not re-started
When a channel transition is still is inTransition() when re-initialized for the next transition, this means it has not received a new target value (because that ends a transition), but OTHER channels in the device are starting a new transition. So we take that as an implicit early abort of the previous transition, and do not init a new one.