Evaluators: state now changes when both conditions ARE true, and one GETS false
Before, only a false-to-true change of either condition could cause an output change.
Now, an already true state will be re-applied when the opposite condition falls away.
This hopefully makes evaluators with overlapping on/off conditions behaving more intuitively.

The old behaviour was technically correct, but without complete history could make the impression of a "blocked" evaluator
- i.e. not changing despite both conditions in a STATE that seems to imply a different output, while the
strict sequence of EVENTS does not.